# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FireGuard AI - Hybrid fire detection system API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type SensorStatus = (typeof SensorStatus)[keyof typeof SensorStatus];

export const SensorStatus = {
  online: "online",
  offline: "offline",
  warning: "warning",
  critical: "critical",
} as const;

export type SensorFireRisk =
  (typeof SensorFireRisk)[keyof typeof SensorFireRisk];

export const SensorFireRisk = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Sensor {
  id: number;
  name: string;
  location: string;
  status: SensorStatus;
  /** Temperature in Celsius */
  temperature: number;
  /** Humidity percentage */
  humidity: number;
  /** Smoke level (0-100) */
  smokeLevel: number;
  /** CO level in ppm */
  coLevel: number;
  lastReading: string;
  fireRisk: SensorFireRisk;
}

export interface DetectRequest {
  sensorId: number;
  temperature: number;
  smokeLevel: number;
  coLevel: number;
}

export type DetectionResultRiskLevel =
  (typeof DetectionResultRiskLevel)[keyof typeof DetectionResultRiskLevel];

export const DetectionResultRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface DetectionResult {
  sensorId: number;
  fire: boolean;
  /** ML confidence score 0-1 */
  confidence: number;
  riskLevel: DetectionResultRiskLevel;
  recommendation: string;
  timestamp: string;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  info: "info",
  warning: "warning",
  critical: "critical",
} as const;

export interface Alert {
  id: number;
  sensorId: number;
  sensorName: string;
  location: string;
  severity: AlertSeverity;
  message: string;
  acknowledged: boolean;
  mlConfidence: number;
  temperature: number;
  createdAt: string;
}

export type DashboardSummarySystemStatus =
  (typeof DashboardSummarySystemStatus)[keyof typeof DashboardSummarySystemStatus];

export const DashboardSummarySystemStatus = {
  normal: "normal",
  elevated: "elevated",
  danger: "danger",
  critical: "critical",
} as const;

export interface DashboardSummary {
  totalSensors: number;
  activeSensors: number;
  criticalAlerts: number;
  resolvedAlerts: number;
  avgTemperature: number;
  systemStatus: DashboardSummarySystemStatus;
  lastScanTime: string;
  mlModelAccuracy: number;
}

export type ActivityEventType =
  (typeof ActivityEventType)[keyof typeof ActivityEventType];

export const ActivityEventType = {
  detection: "detection",
  sensor_update: "sensor_update",
  alert: "alert",
  acknowledgment: "acknowledgment",
  system: "system",
} as const;

export type ActivityEventSeverity =
  (typeof ActivityEventSeverity)[keyof typeof ActivityEventSeverity];

export const ActivityEventSeverity = {
  info: "info",
  warning: "warning",
  critical: "critical",
} as const;

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  message: string;
  sensorName: string;
  severity: ActivityEventSeverity;
  timestamp: string;
}