# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FireGuard AI - Hybrid fire detection system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sensors/{id}:
    get:
      operationId: getSensor
      tags: [sensors]
      summary: Get sensor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sensor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /alerts/{id}/acknowledge:
    post:
      operationId: acknowledgeAlert
      tags: [alerts]
      summary: Acknowledge an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Sensor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [online, offline, warning, critical]
        temperature:
          type: number
          description: Temperature in Celsius
        humidity:
          type: number
          description: Humidity percentage
        smokeLevel:
          type: number
          description: Smoke level (0-100)
        coLevel:
          type: number
          description: CO level in ppm
        lastReading:
          type: string
          format: date-time
        fireRisk:
          type: string
          enum: [low, medium, high, critical]
      required:
        - id
        - name
        - location
        - status
        - temperature
        - humidity
        - smokeLevel
        - coLevel
        - lastReading
        - fireRisk
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        sensorId:
          type: integer
        sensorName:
          type: string
        location:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        message:
          type: string
        acknowledged:
          type: boolean
        mlConfidence:
          type: number
        temperature:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sensorId
        - sensorName
        - location
        - severity
        - message
        - acknowledged
        - mlConfidence
        - temperature
        - createdAt
...