# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FireGuard AI - Hybrid fire detection system API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from "./alertSeverity";

export interface Alert {
  id: number;
  sensorId: number;
  sensorName: string;
  location: string;
  severity: AlertSeverity;
  message: string;
  acknowledged: boolean;
  mlConfidence: number;
  temperature: number;
  createdAt: Date;
}