# lib/api-zod/src/generated/types/sensor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FireGuard AI - Hybrid fire detection system API
 * OpenAPI spec version: 0.1.0
 */
import type { SensorFireRisk } from "./sensorFireRisk";
import type { SensorStatus } from "./sensorStatus";

export interface Sensor {
  id: number;
  name: string;
  location: string;
  status: SensorStatus;
  /** Temperature in Celsius */
  temperature: number;
  /** Humidity percentage */
  humidity: number;
  /** Smoke level (0-100) */
  smokeLevel: number;
  /** CO level in ppm */
  coLevel: number;
  lastReading: Date;
  fireRisk: SensorFireRisk;
}